refactor : 계층 변경으로 인한 변경사항
