+++ src/main/java/com/kakaotech/team14backend/post/application/FindPopularPosts.java
@@ -4,12 +4,12 @@ import com.kakaotech.team14backend.common.RedisKey;
 import com.kakaotech.team14backend.outer.post.dto.GetIncompletePopularPostDTO;
 import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListRequestDTO;
 import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListResponseDTO;
-import com.kakaotech.team14backend.outer.post.mapper.PostMapper;
 import com.kakaotech.team14backend.post.domain.LevelIndexes;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.domain.PostRandomFetcher;
 import com.kakaotech.team14backend.post.domain.RandomIndexes;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
+import com.kakaotech.team14backend.post.mapper.PostMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
